greenpulse/requirements.txt
fastapi
uvicorn
jinja2
mysql-connector-python
paho-mqtt
requests
python-multipart
schedule
